backend/pyproject.toml
[tool.poetry]
name = "jew-pizza"
version = "0.0.1-dev"
description = ""
authors = ["David Cooper <david@jew.pizza>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.8"
Jinja2 = "^3.0.2"
django-environ = "^0.7.0"
gunicorn = "^20.1.0"
psycopg2 = "^2.9.1"
whitenoise = "^5.3.0"
django-npm = "^1.0.0"
django-redis = "^5.0.0"
django-light = "^0.1.0"
django-recurrence = "^1.10.3"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^5.2.0"}
django-widget-tweaks = "^1.4.8"
twilio = "^7.0.0"

[tool.poetry.dev-dependencies]
ipython = "^7.28.0"
django-extensions = "^3.1.3"
flake8 = "^3.9.2"
black = "^21.9b0"
isort = "^5.9.3"
ipdb = "^0.13.9"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
pytest-django = "^4.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py38']
exclude = '/migrations/'
experimental-string-processing = true

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
force_sort_within_sections = true
sections = ['FUTURE', 'STDLIB', 'THIRDPARTY', 'DJANGO', 'DJANGOTHIRDPARTY', 'FIRSTPARTY', 'LOCALFOLDER']
known_django = ['django']
known_djangothirdparty = ['environ', 'phonenumber_field', 'recurrence', 'widget_tweaks']
skip = ['migrations']

[tool.pytest.ini_options]
addopts = '-s --cov=. --cov-report=term-missing'
python_files = 'tests.py'
cache_dir = '/tmp/.pytest_cache'
filterwarnings = [
    # From django-recurrance
    'ignore:django.utils.translation.ugettext_lazy\(\) is deprecated in favor of django.utils.translation.gettext_lazy\(\).:django.utils.deprecation.RemovedInDjango40Warning',
]
